refactor: ApisList + ApisListTable components
